Replace git subcommand with open command
We decided that there are significant problems with having a "run
arbitrary git commands in some other directory".

The two obvious problems are that filename completion will be WRONG (not
that it won't work, so that's even worse), and that relative paths could
potentially be wrong too from what the user might expect.

So, instead we add a command to go to the directory.  Then you can run
your arbitrary git commands