Introduce a new variable in hermit add
This makes it more descriptive, allowing me to be isolated from which
file has the suffix.

The goal is to make it so that the file with the suffix is actually the
complete file, so we can avoid having to rename a file in the case of
redacting a committed file.