Get rid of uuid crate and usage in fs test helpers
Uuid should never have been a normal dependency of hermit, at most it
should have been a dev-dependency.  Also, with the switch to not
cleaning up at the end of tests using this, adding a random element
means... that the clean-up at the start of the next test run is
meaningless. So yeah. I had like 700 of those test dirs in my target
directory after working on all these refactors...